The TotalGuests property in ScanSummary.cs is only used in the view for display (@log.TotalGuests), so changing to long? is safe. The other TotalGuests references are in different models/entities.
Here's a summary of the fix:

Root cause: The stored procedure ProcScanSummary uses SUM(g.NoOfMembers) where NoOfMembers is int. SQL Server's SUM(int) returns int, and when the sum exceeds the int max value (~2.1 billion), it throws "Arithmetic overflow error converting expression to data type int."

+++ EventPro.DAL/Models/ScanSummary.cs
@@ -7,7 +7,7 @@ namespace EventPro.DAL.Models
         public string EventTitle { get; set; }
         public string SystemEventTitle { get; set; }
 
-        public int? TotalGuests { get; set; }
+        public long? TotalGuests { get; set; }
         public int Allowed { get; set; }
         public int Declined { get; set; }
     }
